requirements.txt
pynput==1.7.6
PyQt5==5.15.9
requests==2.28.2
PyJWT==2.6.0
